Add files via upload
The primary aim of this project is to develop a predictive model that accurately evaluates and predicts employee promotions based on various features within the dataset. A significant focus is placed on ensuring the fairness and unbiased nature of the model, particularly with regard to gender, to ensure equal opportunity for promotions.